+++ RELEASE-NOTES.md
@@ -6,6 +6,7 @@
 * #94 dflib-jdbc: SqlSaver to allow wrapping DataFrame updates in custom functions
 * #95 Reusable SqlLoader - allow different parameters for the same loader instance
 * #98 DbFlavor as an object with state and behavior
+* #99 DbFlavor for PostgreSQL
 
 ## Release 0.7